+++ include/engine/ecs/entities/Boid2D.h
@@ -10,9 +10,9 @@
 
 struct Boid2D : public Entity{
     public:
-        float neighborRange = 1.0f;
-        float viewRange = 10.0f;
+        double neighborRange = 1.0f;
+        double viewRange = 10.0f;
         RigidbodyComponent rigidbody;
         SpriteComponent sprite = {0};
-        Boid2D(int id, float x, float y, int z);
+        Boid2D(int id, double x, double y, int z);
 };
