+++ src/engine/ecs/components/physics/rigidbody/PositionComponent.cpp
@@ -4,13 +4,13 @@
 
 #include "PositionComponent.h"
 
-PositionComponent::PositionComponent(float x, float y, int z) {
+PositionComponent::PositionComponent(double x, double y, int z) {
     this->x = x;
     this->y = y;
     this->z = z;
 }
 
-PositionComponent::PositionComponent(float x, float y) {
+PositionComponent::PositionComponent(double x, double y) {
     this->x = x;
     this->y = y;
     this->z = 0;
