+++ src/engine/ecs/entities/Boid2D.cpp
@@ -3,5 +3,5 @@
 //
 
 #include "Boid2D.h"
-Boid2D::Boid2D(int id, float x, float y, int z)
+Boid2D::Boid2D(int id, double x, double y, int z)
  : Entity(id, {x,y,z}), rigidbody({0, {0, 0.01}}) {}
